+++ cax/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'Christopher Tunnell'
 __email__ = 'ctunnell@nikhef.nl'
-__version__ = '5.0.4'
+__version__ = '5.0.12'